!! (Auth AOP) !!

+++ src/main/java/com/chatter/Chatly/service/MemberService.java
@@ -2,8 +2,10 @@ package com.chatter.Chatly.service;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.chatter.Chatly.dto.MemberDto;
 import com.chatter.Chatly.dto.MemberRequestDto;
@@ -18,9 +20,11 @@ import jakarta.transaction.Transactional;
 public class MemberService {
     private final MemberRepository memberRepository;
     private final BCryptPasswordEncoder passwordEncoder;
-    public MemberService(MemberRepository memberRepository, BCryptPasswordEncoder passwordEncoder) {
+    private final AuthService authService;
+    public MemberService(MemberRepository memberRepository, BCryptPasswordEncoder passwordEncoder, AuthService authService) {
         this.memberRepository = memberRepository;
         this.passwordEncoder = passwordEncoder;
+        this.authService = authService;
     }
 
     public List<MemberDto> getAllMembers(){
@@ -50,6 +54,7 @@ public class MemberService {
     }
 
     public MemberDto updateMember(String id, MemberRequestDto dto){
+        if (!isRequester(id)) throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
         Member member = dto.toEntity();
         Member target = memberRepository.findById(id)
         .orElseThrow(() -> new ResourceNotFoundException("Member not found with ID: " + id));
@@ -58,8 +63,14 @@ public class MemberService {
     }
 
     public void deleteMember(String id){
+        if (!isRequester(id)) throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
         Member member = memberRepository.findById(id)
         .orElseThrow(() -> new ResourceNotFoundException("Member not found with ID: " + id));
         memberRepository.delete(member);
+    }
+
+    public boolean isRequester(String targetId){
+        String memberId = authService.getMemberIdFromRequest();
+        return memberId!=null && memberId.equals(targetId);
     }
 }
